% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{Major Asset Class ETF Data}
\format{
\subsection{\code{prices}}{

A data frame with 68,047 rows and 8 columns:
\describe{
\item{ticker}{ETF ticker}
\item{date}{Date in format YYYY-MM-DD}
\item{open}{Day's opening price}
\item{high}{Day's high price}
\item{low}{Day's low price}
\item{close}{Day's closing price}
\item{volume}{Day's traded volume}
\item{closeadjusted}{Day's closing price adjusted for dividends and
corporate actions}
}
}
}
\source{
\url{https://www.robotwealth.com}
}
\usage{
prices
}
\description{
Historical price and volume data for 11 ETFs representing major asset
classes.
}
\keyword{datasets}
