% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r1000}
\alias{r1000}
\title{Russell 1000 Price Data}
\format{
\subsection{\code{r1000}}{

A data frame with  9,438,170 rows and 9 columns:
\describe{
\item{ticker}{Stock ticker}
\item{date}{Date in format YYYY-MM-DD}
\item{open}{Day's opening price}
\item{high}{Day's high price}
\item{low}{Day's low price}
\item{close}{Day's closing price}
\item{unadjusted_close}{Day's closing price adjusted for dividends and
corporate actions}
\item{volume}{Day's traded volume}
\item{is_universe}{Boolean indicating whether stock was in the R1000 on a
particular day}
}
}
}
\source{
\url{https://www.robotwealth.com}
}
\usage{
r1000
}
\description{
Historical price and volume data for Rusell 1000 constituent stocks.
}
\keyword{datasets}
