% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_r1000.R
\name{load_r1000}
\alias{load_r1000}
\title{Load the Russell 1000 Data}
\usage{
load_r1000()
}
\value{
The data is loaded into the global environment. This function does not return a value.
}
\description{
This function loads the Russell 1000 dataset into the global environment.
If the data file is not already available locally, it downloads the file
and saves it to the \code{data} directory before loading it.
}
\details{
The function first checks if the file \code{r1000.rda} exists in the \code{data} directory.
If the file does not exist, it downloads the data using
\code{utils::download.file}. After downloading, the file is loaded into the global environment.
}
\examples{
\dontrun{
load_r1000()
}

}
