% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlaqData-package.R
\docType{package}
\name{tlaqData-package}
\alias{tlaqData}
\alias{tlaqData-package}
\title{tlaqData: Data used in Trade Like a Quant Bootcamp}
\description{
Includes price history for ETFs representing major asset classes and Russell 1000 constituents.
}
\author{
\strong{Maintainer}: Kris Longmore \email{kris@robotwealth.com}

}
\keyword{internal}
